requirements.txt
ipykernel
openai
streamlit
streamlit_webrtc
moviepy
pydub
python-dotenv
requests
pandas
plotly-express